+++ src/contexts/provider/NotificationProvider.tsx
@@ -1,0 +1,38 @@
+import { ReactNode, useState } from "react";
+import { NotificationContextType, Notification } from "../../types";
+import { NotificationContext } from "../NotificationContext";
+
+type Props = {
+  children: ReactNode;
+};
+
+export const NotificationProvider = ({ children }: Props) => {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = (message: string, type: Notification["type"]) => {
+    const newNotification: Notification = {
+      id: Date.now(),
+      message,
+      type,
+    };
+    setNotifications((prev) => [...prev, newNotification]);
+  };
+
+  const removeNotification = (id: number) => {
+    setNotifications((prev) =>
+      prev.filter((notification) => notification.id !== id),
+    );
+  };
+
+  const notificationContext: NotificationContextType = {
+    notifications,
+    addNotification,
+    removeNotification,
+  };
+
+  return (
+    <NotificationContext.Provider value={notificationContext}>
+      {children}
+    </NotificationContext.Provider>
+  );
+};
