+++ src/contexts/provider/ThemeProvider.tsx
@@ -1,0 +1,26 @@
+import { ReactNode, useState } from "react";
+import { ThemeContextType } from "../../types";
+import { ThemeContext } from "../ThemeContext";
+
+type Props = {
+  children: ReactNode;
+};
+
+export const ThemeProvider = ({ children }: Props) => {
+  const [theme, setTheme] = useState("light");
+
+  const toggleTheme = () => {
+    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+  };
+
+  const themeContext: ThemeContextType = {
+    theme,
+    toggleTheme,
+  };
+
+  return (
+    <ThemeContext.Provider value={themeContext}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
